wip: split comps
