+++ src/containers/main/SideBar/Miner/components/ModeSelect.tsx
@@ -1,3 +1,4 @@
+import { CustomPowerLevelsDialog } from '@app/containers/main/SideBar/Miner/components/CustomPowerLevels/CustomPowerLevelsDialog';
 import { ModeSelectWrapper, TileItem } from '../styles';
 import { useTranslation } from 'react-i18next';
 import { Typography } from '@app/components/elements/Typography.tsx';
@@ -11,7 +12,6 @@ import { useAppStateStore } from '@app/store/appStateStore.ts';
 import { useMiningStore } from '@app/store/useMiningStore.ts';
 import { useAppConfigStore } from '@app/store/useAppConfigStore';
 import { modeType } from '@app/store/types';
-import { CustomPowerLevelsDialog } from './CustomPowerLevelsDialog';
 
 function ModeSelect() {
     const { t } = useTranslation('common', { useSuspense: false });
